task-1/scrape_lnkd/requirements.txt
selenium==4.24.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
pandas==2.2.2
python-dotenv==1.0.1
